ia16/DOS experiment: compile fixes + findings (issue #19429)
Builds on agatti's diff. Small model compiles fully but needs 96KiB of
text (64KiB near-code limit); medium model breaks MicroPython's
fnptr-in-void* slot tables (168 "non-enclosed address space" errors).
Full analysis in NOTES-IA16.md.

+++ ports/unix/fatfs_port.c
@@ -4,10 +4,10 @@
 DWORD get_fattime(void) {
     time_t now = time(NULL);
     struct tm *tm = localtime(&now);
-    return ((1900 + tm->tm_year - 1980) << 25)
-           | ((tm->tm_mon + 1) << 21)
-           | (tm->tm_mday << 16)
-           | (tm->tm_hour << 11)
-           | (tm->tm_min << 5)
-           | (tm->tm_sec / 2);
+    return ((1900 + (uint32_t)tm->tm_year - 1980) << 25)
+           | (((uint32_t)tm->tm_mon + 1) << 21)
+           | ((uint32_t)tm->tm_mday << 16)
+           | ((uint32_t)tm->tm_hour << 11)
+           | ((uint32_t)tm->tm_min << 5)
+           | ((uint32_t)tm->tm_sec / 2);
 }
